# layernorm: supplement for adaptive wg size adjustment